Split single-module library
